Sorting task 0 error

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,8 +1,8 @@
 #include "search_algos.h"
 
 /**
- * linear_search - searches for a value in an array
- * using Linear search
+ * linear_search - searches for a value in an array of
+ * integers using the Linear search algorithm
  *
  * @array: pointer to first element of array
  * @size: number of elements in array
@@ -11,14 +11,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	if (array == NULL)
-		return (-1);
+        if (array == NULL)
+                return (-1);
 
-	for (int i = 0; i < (int)size; i++)
-	{
-		printf("Value checked array[%u] = [%d]\n", i, array[i]);
-		if (value == array[i])
-			return (i);
-	}
-	return (-1);
+        for (size_t i = 0; i < size; i++)
+        {
+                printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+                if (value == array[i])
+                        return ((int)i);
+        }
+        return (-1);
 }
